requirements.txt
discord
requests
jsons
python-dotenv